Compile for visionOS
