+++ Sources/StoreHelper/Styles/Shared/Buttons.swift
@@ -16,7 +16,7 @@ public extension View {
                                                pressedColor: Color = .secondary,
                                                padding: EdgeInsets = EdgeInsets(top: 10, leading: 10, bottom: 10, trailing: 10)) -> some View {
         
-        #if os(iOS)
+        #if os(iOS) || os(visionOS)
         self.buttonStyle(.borderedProminent)
         #else
         self.buttonStyle(macOSButtonStyle(foregroundColor: foregroundColor, backgroundColor: backgroundColor, pressedColor: pressedColor, padding: padding))
