+++ Sources/StoreHelper/Styles/macOS/Image-macos.swift
@@ -10,29 +10,6 @@ import SwiftUI
 #if os(macOS)
 @available(macOS 12.0, *)
 public extension Image {
-    
-    // Read images from the Sources/Resources folder
-    init(packageResource name: String, ofType type: String) {
-        #if canImport(UIKit)
-        guard let path = Bundle.module.path(forResource: name, ofType: type),
-              let image = UIImage(contentsOfFile: path) else {
-                  self.init(name)
-                  return
-              }
-        
-        self.init(uiImage: image)
-        #elseif canImport(AppKit)
-        guard let path = Bundle.module.path(forResource: name, ofType: type),
-              let image = NSImage(contentsOfFile: path) else {
-                  self.init(name)
-                  return
-              }
-        self.init(nsImage: image)
-        #else
-        self.init(name)
-        #endif
-    }
-    
     func bodyImage() -> some View {
         self
             .resizable()
@@ -40,12 +17,6 @@ public extension Image {
             .aspectRatio(contentMode: .fit)
             .frame(maxWidth: 1200)
             .padding(EdgeInsets(top: 10, leading: 0, bottom: 20, trailing: 10))
-    }
-    
-    func bodyImageNotRounded() -> some View {
-        self
-            .resizable()
-            .aspectRatio(contentMode: .fit)
     }
     
     func bodyImageConstrained(width: CGFloat? = nil, height: CGFloat? = nil) -> some View {
